Remove incremental until I decide to put it back

+++ txorbit/__init__.py
@@ -1,3 +1,3 @@
 from ._version import __version__ as version
 
-__version__ = version.short()
+__version__ = version
